python-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
guardrails-ai==0.4.2
pydantic==2.5.0
python-multipart==0.0.6 